WRR-21063: Enabled scrolling for Tabs within TabLayout Component (#36)
### Checklist

* [x] I have read and understand the [contribution
guide](https://enactjs.com/docs/developer-guide/contributing/)
* [ ] A [CHANGELOG
entry](https://enactjs.com/docs/developer-guide/contributing/changelogs/)
is included
* [x] At least one test case is included for this feature or bug fix
* [x] Documentation was added or is not needed
* [ ] This is an API breaking change

### Issue Resolved / Feature Added
[//]: # (Describe the issue resolved or feature added by this pull
request)
Enabled scrolling for Tabs within TabLayout Component

### Resolution
[//]: # (Does the code work as intended?)
[//]: # (What is the impact of this change and *why* was it made?)


### Additional Considerations
[//]: # (How should the change be tested?)
[//]: # (Are there any outstanding questions?)
[//]: # (Were any side-effects caused by the change?)
The horizontal line (divider) does not fit the entire content of the
TabGroup. This is because the TabGroup now uses a Scroller component,
and the content that overflows is not covered. However, this issue can
be left as it is, as the horizontal line will be completely removed in
the new design.

### Links
[//]: # (Related issues, references)
WRR-21063

### Comments

Enact-DCO-1.0-Signed-off-by: Ion Andrusciac
ion.andrusciac@lgepartner.com

---------

Co-authored-by: Mikyung Kim <mikyung27.kim@lge.com>