WRR-20094: Changes made to InputField to match the new GUI (#59)
Enact-DCO-1.0-Signed-off-by: Alexandru Morariu
alexandru.morariu@lgepartner.com

### Checklist

* [X] I have read and understand the [contribution
guide](https://enactjs.com/docs/developer-guide/contributing/)
* [x] A [CHANGELOG
entry](https://enactjs.com/docs/developer-guide/contributing/changelogs/)
is included
* [ ] At least one test case is included for this feature or bug fix
* [X] Documentation was added or is not needed
* [ ] This is an API breaking change

### Issue Resolved / Feature Added
[//]: # (Describe the issue resolved or feature added by this pull
request)
I changed InputField to match the new GUI

### Resolution
[//]: # (Does the code work as intended?)
[//]: # (What is the impact of this change and *why* was it made?)


### Additional Considerations
[//]: # (How should the change be tested?)
[//]: # (Are there any outstanding questions?)
[//]: # (Were any side-effects caused by the change?)
I added a new potential CSS class to input field in order to apply
letter spacing to password. Up until now we could only apply it to
passwordtel.

I could not manage to implement 2 of the required features. The first is
to change the size of the caret, the second is to change the color of
the password dot that we are currently affecting.

### Links
[//]: # (Related issues, references)
WRR-20094

### Comments

---------

Co-authored-by: Daniel Stoian <daniel.stoian@lgepartner.com>
Co-authored-by: Daniel Stoian <63335068+daniel-stoian-lgp@users.noreply.github.com>
Co-authored-by: Mikyung Kim <mikyung27.kim@lge.com>