WRR-21578: Fixed TabLayout to match the latest GUI (#93)
### Checklist

* [x] I have read and understand the [contribution
guide](https://enactjs.com/docs/developer-guide/contributing/)
* [ ] A [CHANGELOG
entry](https://enactjs.com/docs/developer-guide/contributing/changelogs/)
is included
* [ ] At least one test case is included for this feature or bug fix
* [x] Documentation was added or is not needed
* [ ] This is an API breaking change

### Issue Resolved / Feature Added
[//]: # (Describe the issue resolved or feature added by this pull
request)
Fixed TabLayout to match the latest GUI guide
Tab Scrolling needs to be implemented as the GUI guide

### Resolution
[//]: # (Does the code work as intended?)
[//]: # (What is the impact of this change and *why* was it made?)
When Tab Scrolling is enabled in horizontal orientation, tabGroup
container bg needs to be expanded to the full screen

### Additional Considerations
[//]: # (How should the change be tested?)
[//]: # (Are there any outstanding questions?)
[//]: # (Were any side-effects caused by the change?)
I think to minimize the code implementing from the app side(sampler
code), we need to add a prop that specifies how much padding should be
added to the horizontal tabs. In the qa-sampler, It was 132px.

### Links
[//]: # (Related issues, references)
WRR-21578

### Comments

---------

Co-authored-by: Mikyung Kim <mikyung27.kim@lge.com>