NXT-550: Adapted screenshot and ui tests for Chrome 132 (#181)
### Checklist

* [x] I have read and understand the [contribution
guide](http://enactjs.com/docs/developer-guide/contributing/)
* [ ] A [CHANGELOG
entry](http://enactjs.com/docs/developer-guide/contributing/changelogs/)
is included
* [ ] At least one test case is included for this feature or bug fix
* [x] I have run automated testing and it is passed
* [x] Documentation was added or is not needed
* [ ] This is an API breaking change

### Issue Resolved / Feature Added
[//]: # (Describe the issue resolved or feature added by this pull
request)
Updated ui tests to work in Chrome 132

### Resolution
[//]: # (Does the code work as intended?)
[//]: # (What is the impact of this change and *why* was it made?)


### Additional Considerations
[//]: # (How should the change be tested?)
[//]: # (Are there any outstanding questions?)
[//]: # (Were any side-effects caused by the change?)

- Updated 2 screenshot tests view to import limestone components from
local path
- ui/PopupTabLayout-specs.js - the first tabItem in the view should be
focused on render, according to our API. I'm not sure why the previous
test logic worked. It must have found another element on the page that
satisfied the criteria and first tabItem was put as second element in
array. Manual check of the logic shows that the tabItem(0) should be
focused
- ScrollerWithEditableSelectItemByLongPress-specs.js - the removed steps
are irrelevant for the purpose of the test. It was focusing a certain
element on the page instead of the first focusable element. The
introduced delays were added as, sometimes, the test logic would not
wait for editable mode to activate
- VirtualGridList-QWTC-2029 - irrelevant logic. it was trying to move
focus out of an inputField in the view and then refocus the same
inputField right away. The inputField is just a helper for the view,
trying to set the data size for virtual list

### Links
[//]: # (Related issues, references)
WRR-30883

### Comments

Enact-DCO-1.0-Signed-off-by: Daniel Stoian
(daniel.stoian@lgepartner.com)

---------

Co-authored-by: taeyoung.hong <taeyoung.hong@lge.com>